server/pyproject.toml
[tool.poetry]
name = "Elucidate API"
version = "0.1.0"
description = ""
authors = ["Coders for Causes <hello@codersforcauses.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^22.1"
psycopg2 = "^2.9.3"
drf-jwt = "^1.19.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
